+++ bff/src/main/java/bg/com/bo/bff/providers/interfaces/IAchAccountProvider.java
@@ -15,7 +15,7 @@ import java.util.Map;
 public interface IAchAccountProvider {
     ClientToken generateAccessToken() throws IOException;
     GenericResponse addAchAccount(String accessToken, AddAchAccountBasicRequest addAchAccountBasicRequest, Map<String, String> parameters) throws IOException;
-    GenericResponse deleteAchAccount(String personId, int identifier, String deviceId, String deviceIp) throws IOException;
+    GenericResponse deleteAchAccount(String personId, long identifier, String deviceId, String deviceIp) throws IOException;
     BanksMWResponse getBanks() throws IOException;
     BranchOfficeMWResponse getAllBranchOfficeBank(Integer code) throws IOException;
     AchAccountMWResponse getAchAccounts(Integer personId, Map<String, String> parameters) throws IOException;
